Fixed bug identified by Holly:
"If you start at D-1485 quick select and then choose User Define Quick Select and then check ON the Export-Inflow Ratio checkbox on the D-1641/D-485 regulations tab, none of the radio buttons on the left-hand side are automatically checked. Please have the D1641 radio button selected in a scenario such as this. This means that a “1” should be written to Index 2 of the GUI_Regs.table. "